Different dataset

+++ train_discrepancy_detection.py
@@ -89,8 +89,8 @@ def train_discrepancy_detection(config):
             ids2 = data['ids2'].to(device, dtype=torch.long)
             mask2 = data['mask2'].to(device, dtype=torch.long)
             token_type_ids2 = data['token_type_ids2'].to(device, dtype=torch.long)
-            #targets = data['targets'].to(device, dtype=torch.float)
-            targets = data['targets'].to(device, dtype=torch.long)
+            targets = data['targets'].to(device, dtype=torch.float)
+            #targets = data['targets'].to(device, dtype=torch.long)
             #targets = nn.functional.one_hot(targets)
 
             #targets = nn.functional.one_hot(targets)
@@ -151,8 +151,8 @@ def train_discrepancy_detection(config):
                 token_type_ids2 = data['token_type_ids2'].to(device, dtype=torch.long)
 
                 #token_type_ids = data['token_type_ids'].to(device, dtype=torch.long)
-                #targets = data['targets'].to(device, dtype=torch.float)
-                targets = data['targets'].to(device, dtype=torch.long)
+                targets = data['targets'].to(device, dtype=torch.float)
+                #targets = data['targets'].to(device, dtype=torch.long)
 
                 #outputs = model(ids, mask, token_type_ids)
                 #outputs = model(ids1, mask1, ids2, mask2)
